move reusable functions to utils
