+++ tests/integration-tests/tests/schedulers/test_slurm.py
@@ -29,6 +29,8 @@ from utils import (
     get_instance_info,
     test_cluster_health_metric,
     wait_for_computefleet_changed,
+    _get_alarm_records,
+    _get_start_end_timestamp,
 )
 
 from tests.common.assertions import (
@@ -328,23 +330,6 @@ def _test_protected_mode_alarm(cw_client, cluster_name):
     _verify_alarms(protected_mode_alarm, "ClusterInProtectedMode", cluster_name)
 
 
-def _get_start_end_timestamp(minutes):
-    """
-    The end time for query will be the current time rounded to minute that is not earlier than the current time (ceil).
-    For instance, if the current time is 09:34:20, then the end time for query will be 09:35:00.
-    This is because our metrics have a period of 1 minute, and according to public documentation of GetMetricData:
-    "For better performance, specify StartTime and EndTime values that align with the value of the metric's Period
-    and sync up with the beginning and end of an hour."
-    Reference: https://docs.aws.amazon.com/AmazonCloudWatch/latest/APIReference/API_GetMetricData.html
-    """
-    now_utc = datetime.datetime.now().astimezone(datetime.timezone.utc)
-    end_timestamp_ceil = math.ceil(now_utc.timestamp() / 60) * 60
-    end_dt = datetime.datetime.fromtimestamp(end_timestamp_ceil)
-    start_dt = end_dt - datetime.timedelta(minutes=minutes)
-    start_timestamp = start_dt.timestamp()
-    return start_timestamp, end_timestamp_ceil
-
-
 def _get_metric_data(cluster_name, cw_client, start_timestamp, end_timestamp):
     metrics_response = cw_client.get_metric_data(
         MetricDataQueries=[
@@ -373,10 +358,6 @@ def _get_metric_data(cluster_name, cw_client, start_timestamp, end_timestamp):
         record["Values"] for record in metrics_response["MetricDataResults"] if record["Id"] == "protected_mode"
     ]
     return metric_values
-
-
-def _get_alarm_records(response, alarm_name):
-    return [alarm for alarm in response["MetricAlarms"] if alarm["AlarmName"] == alarm_name]
 
 
 def _verify_alarms(alarms, metric_name, cluster_name):
